repo/backend/requirements.txt
# HarborOps Python dependencies
# Install with: pip install -r requirements.txt

# ---------------------------------------------------------------------------
# Core Django
# ---------------------------------------------------------------------------
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0

# ---------------------------------------------------------------------------
# Database
# ---------------------------------------------------------------------------
mysqlclient==2.2.4

# ---------------------------------------------------------------------------
# Authentication / Security
# ---------------------------------------------------------------------------
# (Django ships with PBKDF2-SHA256; no extra auth library needed)

# ---------------------------------------------------------------------------
# Encryption
# ---------------------------------------------------------------------------
cryptography==42.0.8

# ---------------------------------------------------------------------------
# Celery + Redis
# ---------------------------------------------------------------------------
celery==5.4.0
redis==5.0.7
django-celery-results==2.5.1
django-celery-beat==2.6.0

# ---------------------------------------------------------------------------
# File handling / spreadsheet imports
# ---------------------------------------------------------------------------
openpyxl==3.1.5          # Excel .xlsx read/write
python-magic==0.4.27     # MIME type detection for uploads

# ---------------------------------------------------------------------------
# Utilities
# ---------------------------------------------------------------------------
python-dotenv==1.0.1     # Load .env in development
Pillow==10.4.0           # Image validation for photo ID uploads

# ---------------------------------------------------------------------------
# Development / Testing
# ---------------------------------------------------------------------------
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.1
coverage==7.6.1
